Split video capture from video display
